Avance del main y arreglo de variables

+++ Src/Integrador/CuentaBancaria.java
@@ -2,8 +2,8 @@ package Integrador;
 
 public class CuentaBancaria {
 
-    float numeroCuenta;
-    float saldo;
+    double numeroCuenta;
+    double saldo;
     String TipoCuenta;
 
 
@@ -14,14 +14,14 @@ public class CuentaBancaria {
         int TipoCuenta = 0;
     }
 
-    public CuentaBancaria(float numeroCuenta, String TipoCuenta){
+    public CuentaBancaria(double numeroCuenta, String TipoCuenta){
 
         this.numeroCuenta = 402400434;
         this.TipoCuenta = "Ahorros";
 
     }
 
-    public CuentaBancaria(float numeroCuenta, float saldo, String TipoCuenta){
+    public CuentaBancaria(double numeroCuenta, double saldo, String TipoCuenta){
 
         this.numeroCuenta = numeroCuenta;
         this.saldo = saldo;
@@ -30,14 +30,14 @@ public class CuentaBancaria {
     }
 
 
-    public float getNumeroCuenta() {
+    public double getNumeroCuenta() {
         return numeroCuenta;
     }
     public void setNumeroCuenta(float numeroCuenta) {
         this.numeroCuenta = numeroCuenta;
     }
 
-    public float getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
     public void setSaldo(float saldo) {
